--- AndreMatias.java
@@ -1,10 +1,0 @@
-public class AndreMatias {
-    public static void sayName() {
-        Student student = new Student("Andre", "Matias");
-        System.out.print("My name is ");
-        student.sayFirstName();
-        System.out.print(" ");
-        student.sayLastName();
-        System.out.print("!");
-    }
-}
