--- AndreRibeiro.java
@@ -1,8 +1,0 @@
-
-public class AndreRibeiro {
-
-    public static void sayName() {
-        Student student = new Student("Andre", "Ribeiro");
-        student.sayFullName();
-    }
-}
