--- AndreRodrigues.java
@@ -1,6 +1,0 @@
-public class AndreRodrigues {
-    public static void sayName() {
-        Student student = new Student("Andre", "Rodrigues");
-        student.sayFullName();
-    }
-}
