--- DanielDias.java
@@ -1,10 +1,0 @@
-public class DanielDias {
-    public static void sayName (){
-        Student student = new Student("Daniel", "Dias");
-        System.out.print("My Name is ");
-        student.sayFirstName();
-        System.out.print(" ");
-        student.sayLastName();
-        System.out.print("!");
-    }
-}
