+++ Dockerfile
@@ -1,0 +1,5 @@
+FROM openjdk:17.0.2
+
+COPY target/students.jar .
+
+ENTRYPOINT ["java","-jar","students.jar"]
