--- FilipeGuerreiro.java
@@ -1,21 +1,0 @@
-public class FilipeGuerreiro {
-    
-    public static void sayName(){
-        
-
-        Student student =  new Student("Filipe", "Guerreiro");
-
-        
-        System.out.print("My name is ");
-        
-        student.sayFirstName();
-
-        System.out.print(" ");
-
-        student.sayLastName();
-
-        System.out.print("!");
-
-    }
-
-}
