--- Main.java
@@ -1,46 +1,0 @@
-public class Main {
-    public static void main(String[] args) {
-
-        System.out.println("*** Hello TEAM ***\n");
-        DanielDias.sayName();
-        System.out.println("");
-        PauloVargas.sayName();
-        System.out.println("");
-        AfonsoAlves.sayName();
-        System.out.println("");
-        AndreMatias.sayName();
-        System.out.println("");
-        FilipaNunes.sayName();
-        System.out.println("");
-        FilipeGuerreiro.sayName();
-        System.out.println("");
-        JoaoQuintaneiro.sayName();
-        System.out.println("");
-        JoaquimLima.sayName();
-        System.out.println("");
-        MarcosCaldeira.sayName();
-        System.out.println("");
-        MariaCaldeira.sayName();
-        System.out.println("");
-        JoaoMarques.sayName();
-        System.out.println("");
-        MartimViana.sayName();
-        System.out.println("");
-        MiguelLopes.sayName();
-        System.out.println("");
-        RicardoCoelho.sayName();
-        System.out.println("");
-        VascoBatista.sayName();
-        System.out.println("");
-        ViniciusLuz.sayName();
-        System.out.println("");
-        AntonioSantos.sayName();
-        System.out.println("");
-        TomasRodrigues.sayName();
-        System.out.println("");
-	IsabelSoares.sayName();
-        System.out.println("");
-        AndreRibeiro.sayName();
-        System.out.println("");
-	}
-}
