--- MarcosCaldeira.java
@@ -1,10 +1,0 @@
-public class MarcosCaldeira {
-	    public static void sayName() {
-		            Student student = new Student("Marcos", "Caldeira");
-			    System.out.print("My name is ");
-			    student.sayFirstName();
-			    System.out.print(" ");
-			    student.sayLastName();
-		            System.out.print("!");	    
-	    }
-}
