--- MariaCaldeira.java
@@ -1,11 +1,0 @@
-public class MariaCaldeira {
-    public static void sayName() {
-        Student student = new Student("Maria", "Caldeira");
-        
-        System.out.print("My name is ");
-        student.sayFirstName();
-        System.out.print(" ");
-        student.sayLastName();
-        System.out.print("!");
-    }
-}
