+++ src/main/java/StudentApp/StudentAppApplication.java
@@ -1,0 +1,13 @@
+package StudentApp;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class StudentAppApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(StudentAppApplication.class, args);
+	}
+
+}
