+++ src/main/java/StudentApp/Students/AfonsoAlves.java
@@ -1,0 +1,13 @@
+package StudentApp.Students;
+import StudentApp.Student;
+
+public class AfonsoAlves {
+    public static void sayName() {
+        Student student = new Student("Afonso", "Alves");
+        System.out.print("My name is ");
+        student.sayFirstName();
+        System.out.print(" ");
+        student.sayLastName();
+        System.out.print("!");
+    }
+}
