+++ src/main/java/StudentApp/Students/AndreMatias.java
@@ -1,0 +1,13 @@
+package StudentApp.Students;
+import StudentApp.Student;
+
+public class AndreMatias {
+    public static void sayName() {
+        Student student = new Student("Andre", "Matias");
+        System.out.print("My name is ");
+        student.sayFirstName();
+        System.out.print(" ");
+        student.sayLastName();
+        System.out.print("!");
+    }
+}
