+++ src/main/java/StudentApp/Students/AndreRibeiro.java
@@ -1,0 +1,10 @@
+package StudentApp.Students;
+import StudentApp.Student;
+
+public class AndreRibeiro {
+
+    public static void sayName() {
+        Student student = new Student("Andre", "Ribeiro");
+        student.sayFullName();
+    }
+}
