+++ src/main/java/StudentApp/Students/AndreRodrigues.java
@@ -1,0 +1,9 @@
+package StudentApp.Students;
+import StudentApp.Student;
+
+public class AndreRodrigues {
+    public static void sayName() {
+        Student student = new Student("Andre", "Rodrigues");
+        student.sayFullName();
+    }
+}
