+++ src/main/java/StudentApp/Students/FilipaNunes.java
@@ -1,0 +1,13 @@
+package StudentApp.Students;
+import StudentApp.Student;
+
+public class FilipaNunes {
+    public static void sayName() {
+        Student student = new Student("Filipa", "Nunes");
+        System.out.print("My name is ");
+        student.sayFirstName();
+        System.out.print(" ");
+        student.sayLastName();
+        System.out.print("!");
+    }
+}
