+++ src/main/java/StudentApp/Students/FilipeGuerreiro.java
@@ -1,0 +1,24 @@
+package StudentApp.Students;
+import StudentApp.Student;
+
+public class FilipeGuerreiro {
+    
+    public static void sayName(){
+        
+
+        Student student =  new Student("Filipe", "Guerreiro");
+
+        
+        System.out.print("My name is ");
+        
+        student.sayFirstName();
+
+        System.out.print(" ");
+
+        student.sayLastName();
+
+        System.out.print("!");
+
+    }
+
+}
