+++ src/main/java/StudentApp/Students/JoaquimLima.java
@@ -1,0 +1,13 @@
+package StudentApp.Students;
+import StudentApp.Student;
+
+public class JoaquimLima {
+    public static void sayName() {
+        Student student = new Student("Joaquim", "Lima");
+        System.out.print("My Name is ");
+        student.sayFirstName();
+        System.out.print(" ");
+        student.sayLastName();
+        System.out.print("!");
+    }
+}
