+++ src/main/java/StudentApp/Students/MariaCaldeira.java
@@ -1,0 +1,14 @@
+package StudentApp.Students;
+import StudentApp.Student;
+
+public class MariaCaldeira {
+    public static void sayName() {
+        Student student = new Student("Maria", "Caldeira");
+        
+        System.out.print("My name is ");
+        student.sayFirstName();
+        System.out.print(" ");
+        student.sayLastName();
+        System.out.print("!");
+    }
+}
