+++ src/main/java/StudentApp/Students/MartimViana.java
@@ -1,0 +1,14 @@
+package StudentApp.Students;
+import StudentApp.Student;
+
+public class MartimViana {
+
+    public static void sayName() {
+        Student student = new Student("Martim", "Viana");
+        System.out.print("My name is ");
+        student.sayFirstName();
+        System.out.print(" ");
+        student.sayLastName();
+        System.out.print("!");
+    }
+}
