+++ src/main/java/StudentApp/Students/MiguelLopes.java
@@ -1,0 +1,15 @@
+package StudentApp.Students;
+import StudentApp.Student;
+
+public class MiguelLopes {
+    public static void sayName() {
+        Student student = new Student("Miguel", "Lopes");
+        System.out.print("My name is ");
+        student.sayFirstName();
+        System.out.print(" ");
+        student.sayLastName();
+        System.out.print("!");
+    }
+
+    
+}
