+++ src/main/java/StudentApp/Students/TomasRodrigues.java
@@ -1,0 +1,14 @@
+package StudentApp.Students;
+import StudentApp.Student;
+
+public class TomasRodrigues {
+
+    public static void sayName() {
+    	Student student = new Student("Tomás", "Rodrigues");
+    	System.out.print("My name is ");
+    	student.sayFirstName();
+    	System.out.print(" ");
+        student.sayLastName();
+        System.out.print("!");
+    }
+}
